faster version

+++ src/strmatch.h
@@ -5,9 +5,10 @@
 #include <cstdint>
 
 namespace strmatch {
-static inline size_t match(std::byte *left, std::byte *right, size_t right_limit) {
+static inline uint32_t match(std::byte *left, std::byte *right,
+                             uint32_t right_limit) {
     assert(left <= right);
-    for (size_t i = 0; i < right_limit; i++) {
+    for (uint32_t i = 0; i < right_limit; i++) {
         if (left[i] != right[i]) {
             return i;
         }
